+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.27.0
 
-  * Add PHP 8.2 polyfill for `json_validate()`
+  * Add PHP 8.3 polyfill for `json_validate()`
   * Fix `IntlDateFormatter::formatObject()` signature
 
 # 1.26.0